Move profile picture utilities to avatarUtils
