+++ app/utils/refreshCachedUserPhoto.js
@@ -1,4 +1,4 @@
-import { getUserPfp } from "./uploadFile";
+import { getUserPfp } from "./avatarUtils";
 
 export async function refreshCachedUserPhoto(uid, updateCachedUserLogin) {
   const photo = await getUserPfp(uid);
